+++ src/common/vec4.rs
@@ -37,8 +37,8 @@ impl<T> Vec4<T> {
 		Self { x, y, z, w }
 	}
 }
-impl NeighborsT for Vec4<i32> {
-	fn neighbors(&self) -> Vec<Vec4<i32>> {
+impl NeighborsT for Vec4<i16> {
+	fn neighbors(&self) -> Vec<Vec4<i16>> {
 		let mut neighbors = Vec::with_capacity(80);
 		for z in self.z - 1..self.z + 2 {
 			for x in self.x - 1..self.x + 2 {
